gpu-requirements.txt
absl-py==0.13.0
astor==0.8.1
atari-py==0.2.6
cached-property==1.5.2
certifi==2021.5.30
cloudpickle==1.6.0
cycler==0.10.0
fire==0.4.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.39.0
gym==0.19.0
h5py==3.4.0
importlib-metadata==4.8.1
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
numpy==1.21.2
opencv-python==4.5.3.56
opt-einsum==3.3.0
pandas==1.3.2
Pillow==8.3.1
protobuf==3.17.3
pygame==2.0.1
pyglet==1.5.19
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
scipy==1.7.1
six==1.16.0
stable-baselines==2.10.2
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
typing-extensions==3.10.0.2
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
